# docs/swagger.yaml
info:
  contact: {}
paths:
  /admin/createProblem:
    post:
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: title
        in: formData
        name: title
        required: true
        type: string
      - description: context
        in: formData
        name: context
        required: true
        type: string
      - description: max_mem
        in: formData
        name: max_mem
        type: integer
      - description: max_runtime
        in: formData
        name: max_runtime
        type: integer
      - description: category_id
        in: formData
        name: category_id
        type: array
      - description: input_case
        in: formData
        name: input_case
        required: true
        type: string
      - description: output_case
        in: formData
        name: output_case
        required: true
        type: string
      responses:
        "200":
          description: '{"code":"200","msg","","data":""}'
          schema:
            type: string
      summary: 创建题目
      tags:
      - 管理员方法
  /ping:
    get:
      responses:
        "200":
          description: '{"code":"200","data":""}'
          schema:
            type: string
      summary: 测试连接
      tags:
      - 公共方法
  /problem/problemDetail:
    get:
      parameters:
      - description: problem_identity
        in: query
        name: problem_identity
        type: string
      responses:
        "200":
          description: '{"code":"200","msg","","data":""}'
          schema:
            type: string
      summary: 查询问题详情
      tags:
      - 公共方法
  /problem/searchProblemList:
    get:
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: size
        in: query
        name: size
        type: integer
      - description: category_identity
        in: query
        name: category_identity
        type: string
      - description: searchData
        in: query
        name: searchData
        type: string
      responses:
        "200":
          description: '{"code":"200","msg","","data":""}'
          schema:
            type: string
      summary: 查询问题列表
      tags:
      - 公共方法
  /submit/searchSubmitList:
    get:
      parameters:
      - description: page
        in: query
        name: page
        type: integer
      - description: size
        in: query
        name: size
        type: integer
      - description: user_identity
        in: query
        name: user_identity
        type: string
      - description: problem_identity
        in: query
        name: problem_identity
        type: string
      - description: status
        in: query
        name: status
        type: integer
      responses:
        "200":
          description: '{"code":"200","msg","","data":""}'
          schema:
            type: string
      summary: 查询提交列表
      tags:
      - 公共方法
  /user/login:
    post:
      parameters:
      - description: user_name
        in: formData
        name: user_name
        type: string
      - description: password
        in: formData
        name: password
        type: string
      responses:
        "200":
          description: '{"code":"200","msg","","data":""}'
          schema:
            type: string
      summary: 用户登录
      tags:
      - 公共方法
  /user/ranking:
    get:
      parameters:
      - description: page
        in: formData
        name: page
        type: string
      - description: size
        in: formData
        name: size
        type: string
      responses:
        "200":
          description: '{"code":"200","msg","","data":""}'
          schema:
            type: string
      summary: 用户排名
      tags:
      - 公共方法
  /user/register:
    post:
      parameters:
      - description: user_name
        in: formData
        name: user_name
        type: string
      - description: password
        in: formData
        name: password
        type: string
      - description: email
        in: formData
        name: email
        type: string
      - description: phone
        in: formData
        name: phone
        type: string
      - description: code
        in: formData
        name: code
        type: string
      responses:
        "200":
          description: '{"code":"200","msg","","data":""}'
          schema:
            type: string
      summary: 注册用户
      tags:
      - 公共方法
  /user/sendCode:
    post:
      parameters:
      - description: email
        in: formData
        name: email
        type: string
      responses:
        "200":
          description: '{"code":"200","msg","","data":""}'
          schema:
            type: string
      summary: 向邮箱发送验证码
      tags:
      - 公共方法
  /user/userDetail:
    get:
      parameters:
      - description: user_identity
        in: query
        name: user_identity
        type: string
      responses:
        "200":
          description: '{"code":"200","msg","","data":""}'
          schema:
            type: string
      summary: 查用户详细信息
      tags:
      - 公共方法
swagger: "2.0"